move sensitivity to tools package
